% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_data_select.r
\name{ah_data_select}
\alias{ah_data_select}
\title{"ah_data_select" - Selects requested Apple Health information from the extracted data frame}
\usage{
ah_data_select(health_data, type_filter = NULL, year_filter = NULL)
}
\arguments{
\item{health_data}{A data frame containing extracted health data, in a format generated by the ah_import_xml.r function}

\item{type_filter}{Parameter to select the element "type" of health data to extract. Named so as the XML file groups elements by "type". See your data frame to see what types are available.}

\item{year_filter}{Optional parameter to select a specific year of data}
}
\value{
type_data A data frame containing the selected type elements
}
\description{
Selects requested Apple Health information from the extracted data frame
}
\note{
Common types to pass to type_filter include:

BodyMassIndex

BodyMass

Height

HeartRate

DistanceWalkingRunning

StepCount
}
\examples{
# health_data <- ah_import_xml("export.xml")
# type_filter <- "BodyMassIndex"
# ah_data_select( health_data , type_filter )
# bmi_data <-ah_data_select( health_data , type_filter )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
