% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_data_summarise.r
\name{ah_data_summarise}
\alias{ah_data_summarise}
\title{"ah_data_summarise" - Summarises requested Apple Health information from the extracted data frame to actually be useful}
\usage{
ah_data_summarise(health_data, type_filter = NULL, summary_period = "hour")
}
\arguments{
\item{health_data}{A data frame containing extracted health data, in a format generated by the ah_import_xml.r function}

\item{type_filter}{Parameter to select the element "type" of health data to extract. Named so as the XML file groups elements by "type". See your data frame to see what types are available.}

\item{summary_period}{A period of time to summarise. Default is hour}
}
\value{
summary_data A data frame containing the selected type elements
}
\description{
Summarises requested Apple Health information from the extracted data frame to actually be useful
}
\note{
Common types to pass to type_filter include:

BodyMassIndex

BodyMass

Height

HeartRate

DistanceWalkingRunning

StepCount
}
\examples{
# PUT EXAMPLES HERE

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
