% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_export_xslx.r
\name{ah_export_xlsx}
\alias{ah_export_xlsx}
\title{"ah_export_xlsx": Writes imported Apple Health dataframe to XLSX format}
\usage{
ah_export_xlsx(health_data, export_filename)
}
\arguments{
\item{health_data}{A data frame containing the data needed to be exported}

\item{export_filename}{A name for saving the file, written in inverted commas}
}
\description{
Writes imported Apple Health dataframe to XLSX format

This is easy with openxslx but thought I'd write a wrapper for me
}
\examples{
# Loads the data
# ah_export_xlsx( health_data , "excel_conversion.xlsx" )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
