% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_import_xml.r
\name{ah_import_xml}
\alias{ah_import_xml}
\title{"ah_import_xml": Import Apple Health data from the exported XML file}
\usage{
ah_import_xml(import_filename)
}
\arguments{
\item{import_filename}{The name of a XML file containing exported Apple Health data, written in inverted commas}
}
\value{
health_data A data frame containing the extracted health data
}
\description{
Imports Apple Health data from its exported XML file into R. Cleans up the data and type identifiers.
}
\examples{
# Loads the data
# ah_import_xml("export.xml")
# Loads the data, and moves it into a data frame
# health_data <- ah_import_xml("export.xml")

}
\references{
http://www.ryanpraski.com/apple-health-data-how-to-export-analyze-visualize-guide/#4

https://gist.github.com/ryanpraski/ba9baee2583cfb1af88ca4ec62311a3d

http://rpubs.com/Ranthony__/visualizing-iphone-health-app-data-in-R

http://www.tdda.info/in-defence-of-xml-exporting-and-analysing-apple-health-data
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
