% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_shiny.r
\name{ah_shiny}
\alias{ah_shiny}
\title{"ah_shiny" - Creates a Shiny dashboard for data analysis}
\usage{
ah_shiny(health_data)
}
\arguments{
\item{health_data}{A data frame containing extracted health data, in a format generated by the ah_import_xml.r function}
}
\description{
Creates a Shiny dashboard for data analysis
}
\examples{
# ah_shiny( health_data )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
