% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_apple_health_xml.r
\name{import_apple_health_xml}
\alias{import_apple_health_xml}
\title{Import Apple Health data from the exported XML file}
\usage{
import_apple_health_xml(filename)
}
\arguments{
\item{filename}{A XML file of the exported Apple Health data}
}
\value{
health_data A data frame containing the extracted health data
}
\description{
Imports Apple Health data from its exported XML file into R. Cleans up the data and type identifiers. Based on R scripts as in the references.
}
\examples{
# Generates two random measurements
measurement1 <- rnorm(100)
measurement2 <- rnorm(100)

}
\references{
(1) http://www.ryanpraski.com/apple-health-data-how-to-export-analyze-visualize-guide/#4

(2) https://gist.github.com/ryanpraski/ba9baee2583cfb1af88ca4ec62311a3d

(3) http://rpubs.com/Ranthony__/visualizing-iphone-health-app-data-in-R
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
