% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah_convert_xml_to_xlsx.r
\name{ah_convert_xml_to_xlsx}
\alias{ah_convert_xml_to_xlsx}
\title{"ah_convert_xml_to_xlsx": Converts Apple Health data from XML to XLSX format}
\usage{
ah_convert_xml_to_xlsx(import_filename, export_filename)
}
\arguments{
\item{import_filename}{The name of a XML file containing exported Apple Health data}

\item{export_filename}{The name of the XLSX output filename}
}
\description{
Converts Apple Health data from XML to XLSX format in one step.

You can easily do this manually but wrote it to save time for people.

This imports parameters from two other functions. This function does nothing with the health_data parameter though.
}
\examples{
# Loads the data
# ah_convert_xml_to_xlsx( "export.xml" , "excel_conversion.xlsx" )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
